Mejoras en precisión de diseño
Se hicieron mejoras en el código para poder hacer cambios precisos en el diseño en cuanto la posición de los objetos.

+++ Source/SynthTabComponent.cpp
@@ -406,7 +406,7 @@ void SynthTabComponent::resized()
     }
 
 
-    auto scaleAndSet = [&](juce::Component& comp, const juce::Rectangle<int>& designRect)
+    auto scaleAndSet = [&](juce::Component& comp, const juce::Rectangle<float>& designRect)
         {
             const float scaledX = guiArea.getX() + designRect.getX() * scale;
             const float scaledY = guiArea.getY() + designRect.getY() * scale + offsetFactor;
